segundaentrega

+++ src/EjerciciosPrimeraEntrega/ejercicioCuatro.java
@@ -2,7 +2,7 @@ package EjerciciosPrimeraEntrega;
 
 public class ejercicioCuatro {
 	    public static void main(String[] args) {
-	        double celsius = 25;
+	        int celsius = 25;
 	        double fahrenheit = (celsius * 9/5) + 32;
 
 	        System.out.println("La temperatura en Fahrenheit es: " + fahrenheit);
